# api/public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenEdusync API",
        "description": "API documentation for the OpenEdusy application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http:\/\/localhost\/api\/v1",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "\/api\/Section": {
            "get": {
                "tags": [
                    "Section"
                ],
                "summary": "Get paginated list of Section",
                "description": "Retrieve a paginated list of Section with optional search",
                "operationId": "getSectionPaginated",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "rows",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/PaginatedSectionResponse200"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Section"
                ],
                "summary": "Create a new Section",
                "description": " Create a new Section with the provided details",
                "operationId": "createSection",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/Section"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Section created successfully",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/CreateSectionResponse200"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/InternalServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "\/api\/Section\/{id}": {
            "get": {
                "tags": [
                    "Section"
                ],
                "summary": "Get a specific Section",
                "description": "Retrieve a Section by its ID",
                "operationId": "getSectionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/GetSectionResponse200"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Section not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Section"
                ],
                "summary": "Update a Section",
                "description": "Update an existing Section with the provided details",
                "operationId": "updateSection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/Section"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Section updated successfully",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/UpdateSectionResponse200"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Section not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/InternalServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Section"
                ],
                "summary": "Delete a Section",
                "description": "Delete a Section by its ID",
                "operationId": "deleteSection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Section deleted successfully",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/DeleteSectionResponse200"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Section not found"
                    },
                    "500": {
                        "description": " Internal server error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/InternalServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}